backend/requirements.txt
alembic==1.14.0
annotated-types==0.7.0
anyio==4.12.1
asyncpg==0.30.0
azure-storage-blob==12.24.0
certifi==2026.2.25
click==8.3.1
fastapi>=0.111.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.27.0
idna==3.11
pydantic==2.7.0
pydantic-settings==2.2.1
pydantic_core==2.18.1
psycopg2-binary==2.9.10
python-chess==1.2.0
chessimg2pos>=0.1.6
python-dotenv==1.0.1
python-jose>=3.4.0
python-multipart>=0.0.22
python-magic==0.4.27
nanoid==2.0.0
slowapi==0.1.9
sniffio==1.3.1
SQLAlchemy[asyncio]==2.0.36
starlette>=0.47.2
typing_extensions==4.15.0
tzdata==2025.3
uvicorn==0.29.0
uvloop==0.22.1
